beenThereBeginning
starting to add function that adds a “Been There” banner when a user
mouses over a marker. idea is to make it so user can update if they’ve
been there, want to go, or haven’t been, maybe add some notes as to how
their experience was, too. should implement knockout for this